+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -230,6 +230,17 @@ def test_pascal_case():
     assert res == "GoodLifeHereV2"
 
 
+def test_capital_case():
+    res = tools.capital_case("PowerShell Remoting")
+    assert res == "PowerShell Remoting"
+    res = tools.capital_case("good life")
+    assert res == "Good Life"
+    res = tools.capital_case("good_life-here v2")
+    assert res == "Good_life-here V2"
+    res = tools.capital_case("")
+    assert res == ""
+
+
 class TestPrintColor:
     def test_print_color(self, mocker):
         mocker.patch('builtins.print')